+++ data/binding/mapbinding.go
@@ -286,6 +286,11 @@ func (b *boundStruct) Reload() (retErr error) {
 		if !f.CanSet() {
 			continue
 		}
+		kind := f.Kind()
+		if kind == reflect.Slice || kind == reflect.Struct {
+			fyne.LogError("Data binding does not yet support slice or struct elements in a struct", nil)
+			continue
+		}
 
 		key := t.Field(j).Name
 		old := (*b.val)[key]
@@ -294,7 +299,7 @@ func (b *boundStruct) Reload() (retErr error) {
 		}
 
 		var err error
-		switch f.Kind() {
+		switch kind {
 		case reflect.Bool:
 			err = b.items[key].(*reflectBool).Set(f.Bool())
 		case reflect.Float32, reflect.Float64: